Switch to double,make ui more controlable
- use double to have finer lock control on programms like vlc
- use numeric up down for easy number editing
- use checkbox for locking

+++ ControlMic/Volume/AudioSetting.cs
@@ -2,5 +2,5 @@
 
 namespace ControlMic.Volume
 {
-    public record AudioSetting(string Id, int Volume, bool IsLocked);
+    public record AudioSetting(string Id, double Volume, bool IsLocked);
 }
